add Sieve

+++ src/main/scala/com/svitovyda/PrimeAtNPlaceRunner.scala
@@ -11,12 +11,12 @@ object PrimeAtNPlaceRunner {
     Key.verbose -> true
   ) withWarmer new Warmer.Default
 
-  @volatile var seqStrResult: Long = 0
-  @volatile var seqRecResult: Long = 0
-  @volatile var parResult: Long = 0
+  @volatile var seqStrResult: Int = 0
+  @volatile var seqRecResult: Int = 0
+  @volatile var parResult: Int = 0
 
   def main(args: Array[String]): Unit = {
-    val place: Long = 20000
+    val place: Int = 20000
 
     val seqRecTime = standardConfig measure {
       seqRecResult = PrimeAtNPlace.sequentialReq(place)
